adb: host: Fix windows build after 5c0999c
* Since device_fd is checked against < 0 above,
  it is safe to cast it to uint64_t and fix
  `comparison of integers of different signs`
  error when crosscompiling adb for Windows.

+++ adb/client/commandline.cpp
@@ -911,7 +911,7 @@ static int adb_sideload_install(const char* filename, bool rescue_mode) {
         fd_set fds;
         struct timeval tv;
         FD_ZERO(&fds);
-        FD_SET(device_fd, &fds);
+        FD_SET(static_cast<uint64_t>(device_fd), &fds);
         tv.tv_sec = 1;
         tv.tv_usec = 0;
         int rc = select(device_fd+1, &fds, NULL, NULL, &tv);
